+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Types of changes:
 
 ### Improved / Modified
 - Removed legacy `pkg_resources` logic for loading entry points (`qbraid._entrypoints`), as support for Python 3.9 has been dropped and the project now requires Python 3.10 or higher. ([#1002](https://github.com/qBraid/qBraid/issues/1002))
+- Populated basis gates property in profile of AWS Braket provider device ([#1002](https://github.com/qBraid/qBraid/pull/1002))
 
 ### Deprecated